Remove version from src and address duplicate functions
