+++ src/index.js
@@ -1,18 +1,15 @@
 import alg from './alg'
 import Graph from './graph'
 import json from './json'
-import version from './version'
 
 export {
   alg,
   Graph,
-  json,
-  version
+  json
 }
 
 export default {
   alg,
   Graph,
-  json,
-  version
+  json
 }
