--- tests/version-test.js
@@ -1,8 +1,0 @@
-var expect = require('./chai').expect
-
-describe('version', function () {
-  it('should match the version from package.json', function () {
-    var packageVersion = require('../package').version
-    expect(require('../lib/version')).to.equal(packageVersion)
-  })
-})
